Debugging FE time integration.
Have a working (and converging) implementation of FE with _one_ time
level. Archiving for now as clean up of debugging changes occur.

Note: them __RMSE__ of the numerical solution converges, but L2 norm
doesn't seem to converge. The idea that there was "bug" to do with
the convergence was based on the L2 norm values, so maybe that was
off base.

Also added the `Revise` julia package to the project environment which
really speeds up the development process.